Completed input N chunks, tested and it worked. All functions have been implimented succesfully.

+++ options.h
@@ -3,6 +3,6 @@
 
 #include <stdio.h>
 
-int process_options(int argc, char **argv, long long *nbytes, long long *customsize, int *inputOptions, int *outputOptions);
+int process_options(int argc, char **argv, long long *nbytes, long *customsize, int *inputOptions, int *outputOptions);
 
 #endif
